+++ src/com/ucm/dasi/catan/game/CatanGameEngine.java
@@ -1,27 +1,35 @@
 package com.ucm.dasi.catan.game;
 
 import com.ucm.dasi.catan.board.ICatanEditableBoard;
+import com.ucm.dasi.catan.board.connection.BoardConnection;
+import com.ucm.dasi.catan.board.element.BoardElement;
+import com.ucm.dasi.catan.board.exception.InvalidBoardElementException;
+import com.ucm.dasi.catan.board.structure.BoardStructure;
 import com.ucm.dasi.catan.exception.NonNullInputException;
 import com.ucm.dasi.catan.exception.NonVoidCollectionException;
 import com.ucm.dasi.catan.game.handler.GameEngineHandlersMap;
 import com.ucm.dasi.catan.game.handler.IGameEngineHandlersMap;
 import com.ucm.dasi.catan.player.IPlayer;
+import com.ucm.dasi.catan.request.IBuildConnectionRequest;
+import com.ucm.dasi.catan.request.IBuildStructureRequest;
 import com.ucm.dasi.catan.request.IRequest;
+import com.ucm.dasi.catan.warehouse.exception.NotEnoughtResourcesException;
 
 import java.util.function.Consumer;
-import java.util.function.Function;
 
 public abstract class CatanGameEngine extends CatanGame<ICatanEditableBoard> implements ICatanGameEngine {
 
+    private Consumer<IRequest> errorHandler;
+
     private IGameEngineHandlersMap handlersMap;
 
-    public CatanGameEngine(ICatanEditableBoard board, IPlayer[] players,
-	    Function<CatanGameEngine, IGameEngineHandlersMap> handlersMapGenerator)
+    public CatanGameEngine(ICatanEditableBoard board, IPlayer[] players, Consumer<IRequest> errorHandler)
 	    throws NonNullInputException, NonVoidCollectionException {
-	
+
 	super(board, players);
 
-	this.handlersMap = new GameEngineHandlersMap(handlersMapGenerator.apply(this));
+	this.errorHandler = errorHandler;
+	this.handlersMap = new GameEngineHandlersMap(generateMap());
     }
 
     @Override
@@ -31,8 +39,10 @@ public abstract class CatanGameEngine extends CatanGame<ICatanEditableBoard> imp
 	}
     }
 
-    protected abstract void handleRequestError(IRequest request);
-    
+    protected void handleRequestError(IRequest request) {
+	errorHandler.accept(request);
+    }
+
     private void processTurnRequest(IRequest request) {
 	Class<?>[] interfaces = request.getClass().getInterfaces();
 
@@ -44,10 +54,42 @@ public abstract class CatanGameEngine extends CatanGame<ICatanEditableBoard> imp
 	    }
 	}
     }
-    
+
     @SuppressWarnings("unchecked")
     private <T extends IRequest> void consumeRequest(Consumer<T> consumer, IRequest request) {
-	consumer.accept((T)request);
+	consumer.accept((T) request);
+    }
+
+    private IGameEngineHandlersMap generateMap() {
+	IGameEngineHandlersMap map = new GameEngineHandlersMap();
+
+	map.put(IBuildConnectionRequest.class,
+		(IBuildConnectionRequest request) -> handleBuildConnectionRequest(request));
+	map.put(IBuildStructureRequest.class,
+		(IBuildStructureRequest request) -> handleBuildStructureRequest(request));
+
+	return map;
+    }
+
+    private void handleBuildConnectionRequest(IBuildConnectionRequest request) {
+	BoardConnection element = new BoardConnection(request.getPlayer(), request.getType());
+
+	try {
+	    getBoard().build(element, request.getX(), request.getY());
+	    getActivePlayer().getWarehouse().substract(element.getCost());
+	} catch (InvalidBoardElementException | NotEnoughtResourcesException e) {
+	    handleRequestError(request);
+	}
+    }
+
+    private void handleBuildStructureRequest(IBuildStructureRequest request) {
+	BoardElement element = new BoardStructure(request.getPlayer(), request.getType());
+
+	try {
+	    getBoard().build(element, request.getX(), request.getY());
+	} catch (InvalidBoardElementException e) {
+	    handleRequestError(request);
+	}
     }
 
 }
